feat: @client alias, fix eslint import resolver, refactor Content
