+++ .eslintrc.json
@@ -26,5 +26,10 @@
         "ignoreRestSiblings": true
       }
     ]
+  },
+  "settings": {
+    "import/resolver": {
+      "typescript": {} // this loads <rootdir>/tsconfig.json to eslint
+    }
   }
 }
