+++ package.json
@@ -39,6 +39,7 @@
     "autoprefixer": "^10.4.20",
     "electron": "31.3.1",
     "eslint": "^8.57.0",
+    "eslint-import-resolver-typescript": "^3.6.3",
     "eslint-plugin-import": "^2.29.1",
     "postcss": "^8.4.41",
     "tailwindcss": "^3.4.9",
