+++ tsconfig.json
@@ -9,7 +9,8 @@
     "sourceMap": true,
     "baseUrl": ".",
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@client/*": ["./src/client/*"]
     },
     "outDir": "dist",
     "moduleResolution": "node",
